rename files, add copyright headers, add small comment
